Add Orders migration.

+++ LabFortyMS/LabFortyMS.Orders/Services/Model/OrderUpdateRequestModel.cs
@@ -2,8 +2,8 @@
 {
     public class OrderUpdateRequestModel
     {
-        public double? PriceBuy { get; set; }
+        public decimal? PriceBuy { get; set; }
 
-        public double? PriceSell { get; set; }
+        public decimal? PriceSell { get; set; }
     }
 }
